portar completar_codigos a tmdb.py y tvdb.py
